tests/submission/Microsoft/forms/formtarget.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>formtarget on button element</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta name="description" content="formtarget on button element" />
    </head>
    <body>
        <script src="/resources/testharness.js" ></script>
        <script src="/resources/testharnessreport.js" ></script>

        <div id=log></div>    
        <button style = "display : none" type="submit" formtarget="_self">Submit</button>
        <input style = "display : none" type="submit" formtarget="_self" value="submit"></input>
        <input style = "display : none" type="submit" value="submit"></input>
        
        <script type="text/javascript">
            test(function() {assert_equals(document.getElementsByTagName("button")[0].formTarget, "_self")}, "formtarget on button element");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].formTarget, "_self")}, "formtarget on input element");
            
            var testElem = document.getElementsByTagName("input")[0];
            testElem.formTarget = "_blank";

            test(function() {assert_equals(document.getElementsByTagName("input")[0].formTarget, "_blank")}, "setting formtarget on input element");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].getAttribute("formtarget"), "_blank")}, "setting formtarget will overwrite attribute value");
            
            test(function() {assert_equals(document.getElementsByTagName("input")[1].formTarget, "")}, "formtarget property default value is an empty string");
            
        </script>
    </body>
</html>
